requirements.txt
boto3==1.35.70
botocore==1.35.70
python-dotenv==1.0.1
Requests==2.32.3
numpy==1.25.0
pandas==1.5.3
mypy==1.13.0
psycopg2-binary==2.9.10